+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 All notable changes to Pug will be documented in this file (beginning with v0.1.1 😅).
 
 ## [Unreleased]
+### Fixed
+- Sanitize command strings